second commit

+++ server1/controllers/taskController.js
@@ -1,5 +1,6 @@
 const Task = require("../models/taskModel");
 
+// Controller to create new Record
 const createTask = async (req, res) => {
   try {
     const { task } = req.body;
@@ -17,27 +18,29 @@ const createTask = async (req, res) => {
       });
     }
     // CREATE RECORD IF NEW RECORD
-    const data = await Task.create({ task });
-    res.status(201).json(data);
+    const newTask = await Task.create({ task });
+    res.status(201).json(newTask);
   } catch (error) {
     res.status(404).json({ message: error.message });
   }
 };
 
+// Controller to get records from DB
 const getTask = async (req, res) => {
   try {
-    const data = await (req.params.id ? Task.findById(req.params.id) : Task.find());
-    if (data.length === 0 || data === null) {
+    const tasks = await (req.params.id ? Task.findById(req.params.id) : Task.find());
+    if (tasks === null || tasks.length === 0) {
       return res.status(404).json({
         message: "No Records found!",
       });
     }
-    res.json(data);
+    res.json(tasks);
   } catch (error) {
     res.status(404).json({ message: error.message });
   }
 };
 
+// Controller to update record available in DB
 const updateTask = async (req, res) => {
   try {
     const id = req.params.id;
@@ -48,9 +51,9 @@ const updateTask = async (req, res) => {
         message: `Please enter${!id && " id"} parameter!`,
       });
     }
-    // GET THE RECORD
-    const data = await Task.findOne({ _id: id });
-    if (data === null) {
+    // CHECK IF THE RECORD EXISTS
+    const taskExists = await Task.findOne({ _id: id });
+    if (taskExists === null) {
       return res.status(404).json({
         message: "No Records found!",
       });
@@ -61,7 +64,7 @@ const updateTask = async (req, res) => {
         message: `Please enter valid Status fields!`,
       });
     }
-
+    // UPDATE THE RECORD
     const updatedData = await Task.findByIdAndUpdate(id, { status });
     res.json({ message: "Task Updated!" });
   } catch (error) {
@@ -69,16 +72,23 @@ const updateTask = async (req, res) => {
   }
 };
 
+// Controller to delete records from DB
 const deleteTask = async (req, res) => {
   try {
+    const id = req.params.id;
+    if (!id) {
+      return res.status(404).json({
+        message: `Please enter valid${!id && " id"} parameter!`,
+      });
+    }
+    // FIND IF THE RECORDS EXISTS
     const task = await Task.findById({ _id: req.params.id });
-
     if (!task) {
       return res.status(404).json({
         message: "No such Task found!",
       });
     }
-
+    // DELETE THE CONTACT BASED IN 'id'
     const deletedtask = await Task.findByIdAndRemove(req.params.id);
     return res.json({ message: "Task Deleted!" });
   } catch (error) {
